requirements.txt
jinja2
jsonschema
mergedeep
requests
pyyaml
validators
coloredlogs
